requirements.txt
Flask==3.1.1
google-api-python-client==2.169.0
google-auth==2.40.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
requests==2.32.3
python-dotenv==1.1.0
gunicorn